Revert "Merge branch 'master' into Ice"
This reverts commit bffddc54367bcce96ffb60e446b4000035306259, reversing
changes made to 286fd8d522b59f6e73b42e4c4deb6502f512fb73.
