+++ AIM_UNPREDICTABLE.cs
@@ -2,16 +2,15 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class HITPOINT : MonoBehaviour {
-    private float t;
+public class AIM_UNPREDICTABLE : MonoBehaviour {
+
 	// Use this for initialization
 	void Start () {
-        t = Time.time;
+		
 	}
 	
 	// Update is called once per frame
 	void Update () {
-        if (Time.time - t > 1)
-            Destroy(gameObject);
+		
 	}
 }
