+++ Scenes/World 2 -- Desert/2-1 -- Windy.unity.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 7e6c545bd1c6a624e9917be3fbd938ee
-folderAsset: yes
-timeCreated: 1534471999
+guid: ddffdf8643c46654dbdf04a61afb9bfe
+timeCreated: 1529293010
 licenseType: Free
 DefaultImporter:
   externalObjects: {}
