--- Scenes/World 2 -- Desert/2-1--Windy_SM.unity.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 00ae388efef56ed498212287a2cf0999
-timeCreated: 1534482878
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
