--- Scenes/World 2 -- Desert/2-2 -- Updraft Downdraft_SM.unity.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 31b6541d9390115438c8575bb883ebdd
-timeCreated: 1535259909
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
