--- Scenes/World 2 -- Desert/2-3 -- Dust Devil_SM.unity.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 5e582386f8938f047b6952da975799f6
-timeCreated: 1535264372
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
