+++ Scenes/World 2 -- Desert/2-4 -- Hot n Cold.unity.meta
@@ -1,10 +1,9 @@
 fileFormatVersion: 2
-guid: dbe990fa88763dd4fb46d4a6626d5731
-timeCreated: 1534472004
+guid: 664ef7b795ae0a746bb0d6f61081eae3
+timeCreated: 1531255437
 licenseType: Free
-NativeFormatImporter:
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 100100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
