--- Script/World 2 -- Desert/AIM_BULLETSPEED.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 613c065a3873aa14d9544835320b144d
-timeCreated: 1535266047
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
