--- Script/World 2 -- Desert/AIM_DESERT1.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: ec54cfb5bc497a24f9fa58a2ea52a24e
-timeCreated: 1534484054
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
