--- Script/World 2 -- Desert/AIM_DESERT2.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 746bd2728d5d6bc4ab509866ac1777e7
-timeCreated: 1534483969
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
