--- Script/World 2 -- Desert/ControllerP1_AITEST_DESERT_HotNCold.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 2c630df57249fa64cbcc75bda1215dfd
-timeCreated: 1535265188
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
