--- Script/World 2 -- Desert/ControllerP1_AITEST_DESERT_UPDRAFT.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: f796e8d67abc7f342b25ee308c1f38f4
-timeCreated: 1534522099
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
