--- Script/World 2 -- Desert/HITPOINT.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: a89aeeb1ce79a624f94b196185b1afbe
-timeCreated: 1534471683
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
