--- Script/World 4 -- Antarctic/4-2 -- Slip 'n Slide/ControllerP2_4_2.cs.meta~HEAD
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 4bf58423005f03541b8c4f695ec85df1
-timeCreated: 1534559725
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
