+++ Script/World 6 -- Tower/ControllerP1_AITEST_ELEVTOR.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using EZCameraShake;
 
 public class ControllerP1_AITEST_ELEVTOR : MonoBehaviour {
