+++ Script/World 6 -- Tower/ControllerP1_AITEST_TOWER_AC.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using EZCameraShake;
 
 public class ControllerP1_AITEST_TOWER_AC : MonoBehaviour {
